# Conflicts:
#	waspc/src/Wasp/Project/Db/Dev/Postgres.hs

+++ waspc/cli/src/Wasp/Cli/Command/Doctor.hs
@@ -52,7 +52,7 @@ checks =
     ("Docker", checkDocker >> return "running"),
     makePortCheck "Client" WebApp.defaultClientPort,
     makePortCheck "Server" Server.defaultServerPort,
-    makePortCheck "Dev database" Dev.Postgres.defaultDevPort
+    makePortCheck "Dev database" (fromIntegral Dev.Postgres.defaultDevPort)
   ]
   where
     makePortCheck name port =
